requirements.txt
langchain
langchain-openai
langchain-community
requests
beautifulsoup4
lxml
python-dotenv
pydantic
reportlab
Pillow
openai
httpx
PyPDF2
pdfplumber
rich
python-docx
